# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: Swagger Turbo-Carnival - OpenAPI 3.0
  version: 1.0.1
servers:
  - url: https://192.168.0.123:1010/
paths:
  /balance:
    get:
      summary: Получение баланса
      description: Получение баланса пользователя по ID
      operationId: GetBalance
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ID'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Balance'
        '400':
          description: Bad request
  /replenish:
    put:
      summary: Пополнение баланса
      description: Пополнение баланса пользователя по ID на Count единиц
      operationId: ReplenishBalance
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IDAndCount'
        required: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Bad request
  /reserve:
    put:
      summary: Резервирование средств
      description: Списание средств со счета пользователя и зачисление их на резервный счет
      operationId: Reserve
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FullStruct'
        required: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Bad request
  /revenue:
    put:
      summary: Признание выручки
      description: Списание средств с резервного счета
      operationId: Revenue
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FullStruct'
        required: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Bad request

components:
  schemas:
    ID:
      type: object
      properties:
        Id:
          type: integer
          format: uint
          example: 10
    Balance:
      type: object
      properties:
        Id:
          type: integer
          example: 10
        Cash:
          type: integer
          example: 100
    IDAndCount:
      type: object
      properties:
        Id:
          type: integer
          format: uint
          example: 10
        Count:
          type: integer
          format: uint
          example: 50
    FullStruct:
      type: object
      properties:
        Id:
          type: integer
          format: uint
          example: 10
        Service_ID:
          type: integer
          format: uint
          example: 50
        Order_id:
          type: integer
          format: uint
          example: 50
        Count:
          type: integer
          format: uint
          example: 50